add NOPs and waste memory for stressing GPU
to keep the GPU busy as much as possible, "NOP" instructions have been
added, e.g. stressing for loop.

On the memory side, the program allocates the GPU memory with a given
expansion factor (default is 4) with respect to the host memory (CPU)
containing useful data. This means that if the host useful data is 1MB,
4*1MB of memory will be reserved on the GPU side. The computation is
still performed on 1MB of GPU memory, resulting in 3MB of GPU memory
being wasted.

This is needed when we need to stress the GPU memory allocation, but we
do not have enough host memory to launch the kernel on.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>